특정 기간의 매출액은 보통 int 타입의 범위를 넘어가므로 bigint로 변경

+++ db/init_postgresql.py
@@ -45,7 +45,7 @@ create_table_queries = [
             """
                 CREATE TABLE order_sales (
                     date TIMESTAMP,
-                    amount INT
+                    amount BIGINT
                 ) 
             """,
             """
